requirements.txt
torch~=2.0.1
numpy~=1.23.5
lightning~=2.0.2
hydra-core